Description
This is the basic layout, more like a start out template for any website. When you look at the code for any website this is the basic structure you will find. Type CTRL + U if you're on a WINDOWS computer and you will see a bunch of gibberish shown to you. Try to identify this structure on any website of choice. 
1. Visit any website
2. Type CTRL + U 
3. Try and identify this structure.